+++ src/java/org/apache/cassandra/gms/FailureDetector.java
@@ -29,9 +29,6 @@ import javax.management.openmbean.*;
 
 import com.google.common.collect.ImmutableMap;
 
-import com.codahale.metrics.Reservoir;
-import com.codahale.metrics.Snapshot;
-import com.codahale.metrics.UniformSnapshot;
 import com.palantir.cassandra.db.BootstrappingSafetyException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -297,7 +294,7 @@ public class FailureDetector implements IFailureDetector, FailureDetectorMBean
         {
             StorageService.instance.unsafeDisableNode();
             logger.error("Detected local pause longer than Gossiper failed bootstrap timeout (nanos) {}"
-                   + "whilst node was bootstrapping", MAX_BOOTSTRAPPING_NODE_PAUSE_IN_NANOS);
+                         + "whilst node was bootstrapping", MAX_BOOTSTRAPPING_NODE_PAUSE_IN_NANOS);
             StorageService.instance.recordNonTransientError(StorageServiceMBean.NonTransientError.BOOTSTRAP_ERROR,
                                                             ImmutableMap.of("timeoutDuringBootstrap", "true"));
             throw new BootstrappingSafetyException("Bootstrap failed due to gossip timeout");
@@ -394,118 +391,12 @@ public class FailureDetector implements IFailureDetector, FailureDetectorMBean
         sb.append("-----------------------------------------------------------------------");
         return sb.toString();
     }
-
-    public static class ArrivalWindow implements Reservoir
-    {
-        private static final Logger logger = LoggerFactory.getLogger(ArrivalWindow.class);
-        private long tLast = 0L;
-        private final ArrayBackedBoundedStats arrivalIntervals;
-        private double lastReportedPhi = Double.MIN_VALUE;
-
-        // in the event of a long partition, never record an interval longer than the rpc timeout,
-        // since if a host is regularly experiencing connectivity problems lasting this long we'd
-        // rather mark it down quickly instead of adapting
-        // this value defaults to the same initial value the FD is seeded with
-        private final long MAX_INTERVAL_IN_NANO = getMaxInterval();
-
-        ArrivalWindow(int size)
-        {
-            arrivalIntervals = new ArrayBackedBoundedStats(size);
-        }
-
-        private static long getMaxInterval()
-        {
-            String newvalue = System.getProperty("cassandra.fd_max_interval_ms");
-            if (newvalue == null)
-            {
-                return FailureDetector.INITIAL_VALUE_NANOS;
-            }
-            else
-            {
-                logger.info("Overriding FD MAX_INTERVAL to {}ms", newvalue);
-                return TimeUnit.NANOSECONDS.convert(Integer.parseInt(newvalue), TimeUnit.MILLISECONDS);
-            }
-        }
-
-        synchronized void add(long value, InetAddress ep)
-        {
-            assert tLast >= 0;
-            if (tLast > 0L)
-            {
-                long interArrivalTime = (value - tLast);
-                if (interArrivalTime <= MAX_INTERVAL_IN_NANO)
-                {
-                    arrivalIntervals.add(interArrivalTime);
-                    logger.trace("Reporting interval time of {} for {}", interArrivalTime, ep);
-                }
-                else
-                {
-                    logger.debug("Ignoring interval time of {} for {}", interArrivalTime, ep);
-                }
-            }
-            else
-            {
-                // We use a very large initial interval since the "right" average depends on the cluster size
-                // and it's better to err high (false negatives, which will be corrected by waiting a bit longer)
-                // than low (false positives, which cause "flapping").
-                arrivalIntervals.add(FailureDetector.INITIAL_VALUE_NANOS);
-            }
-            tLast = value;
-        }
-
-        double mean()
-        {
-            return arrivalIntervals.mean();
-        }
-
-        // see CASSANDRA-2597 for an explanation of the math at work here.
-        double phi(long tnow)
-        {
-            assert arrivalIntervals.mean() > 0 && tLast > 0; // should not be called before any samples arrive
-            long t = tnow - tLast;
-            lastReportedPhi = t / mean();
-            return lastReportedPhi;
-        }
-
-        public double getLastReportedPhi()
-        {
-            return lastReportedPhi;
-        }
-
-        public String toString()
-        {
-            return Arrays.toString(arrivalIntervals.getArrivalIntervals());
-        }
-
-        @Override
-        public int size()
-        {
-            return arrivalIntervals.size();
-        }
-
-        @Override
-        public void update(long interval)
-        {
-            arrivalIntervals.update(interval);
-        }
-
-        @Override
-        public Snapshot getSnapshot()
-        {
-            return arrivalIntervals.getSnapshot();
-        }
-
-        public long getLastInterval()
-        {
-            return arrivalIntervals.getLastInterval();
-        }
-    }
 }
 
 /*
- This class is not thread safe.
- */
-class ArrayBackedBoundedStats implements Reservoir
+This class is not thread safe.
+*/
+class ArrayBackedBoundedStats
 {
     private final long[] arrivalIntervals;
     private long sum = 0;
@@ -520,22 +411,21 @@ class ArrayBackedBoundedStats implements Reservoir
 
     public void add(long interval)
     {
-        if(index == arrivalIntervals.length)
+        if (index == arrivalIntervals.length)
         {
             isFilled = true;
             index = 0;
         }
 
-        if(isFilled)
+        if (isFilled)
             sum = sum - arrivalIntervals[index];
 
         arrivalIntervals[index++] = interval;
         sum += interval;
-        mean = (double)sum / size();
+        mean = (double) sum / size();
     }
 
-    @Override
-    public int size()
+    private int size()
     {
         return isFilled ? arrivalIntervals.length : index;
     }
@@ -550,30 +440,96 @@ class ArrayBackedBoundedStats implements Reservoir
         return arrivalIntervals;
     }
 
-    @Override
-    public void update(long interval)
-    {
-        add(interval);
-    }
-
-    @Override
-    public Snapshot getSnapshot()
-    {
-        // Based on https://github.com/dropwizard/metrics/blob/release/4.2.x/metrics-core/src/main/java/com/codahale/metrics/SlidingWindowReservoir.java#L35
-        long[] values = new long[size()];
-        for (int i = 0; i < values.length; i++)
-        {
-            synchronized (this)
-            {
-                values[i] = arrivalIntervals[i];
-            }
-        }
-        return new UniformSnapshot(values);
-    }
-
     public long getLastInterval()
     {
         return arrivalIntervals[index % arrivalIntervals.length];
     }
 }
 
+class ArrivalWindow
+{
+    private static final Logger logger = LoggerFactory.getLogger(ArrivalWindow.class);
+    private long tLast = 0L;
+    private final ArrayBackedBoundedStats arrivalIntervals;
+    private double lastReportedPhi = Double.MIN_VALUE;
+
+    // in the event of a long partition, never record an interval longer than the rpc timeout,
+    // since if a host is regularly experiencing connectivity problems lasting this long we'd
+    // rather mark it down quickly instead of adapting
+    // this value defaults to the same initial value the FD is seeded with
+    private final long MAX_INTERVAL_IN_NANO = getMaxInterval();
+
+    ArrivalWindow(int size)
+    {
+        arrivalIntervals = new ArrayBackedBoundedStats(size);
+    }
+
+    private static long getMaxInterval()
+    {
+        String newvalue = System.getProperty("cassandra.fd_max_interval_ms");
+        if (newvalue == null)
+        {
+            return FailureDetector.INITIAL_VALUE_NANOS;
+        }
+        else
+        {
+            logger.info("Overriding FD MAX_INTERVAL to {}ms", newvalue);
+            return TimeUnit.NANOSECONDS.convert(Integer.parseInt(newvalue), TimeUnit.MILLISECONDS);
+        }
+    }
+
+    synchronized void add(long value, InetAddress ep)
+    {
+        assert tLast >= 0;
+        if (tLast > 0L)
+        {
+            long interArrivalTime = (value - tLast);
+            if (interArrivalTime <= MAX_INTERVAL_IN_NANO)
+            {
+                arrivalIntervals.add(interArrivalTime);
+                logger.trace("Reporting interval time of {} for {}", interArrivalTime, ep);
+            }
+            else
+            {
+                logger.debug("Ignoring interval time of {} for {}", interArrivalTime, ep);
+            }
+        }
+        else
+        {
+            // We use a very large initial interval since the "right" average depends on the cluster size
+            // and it's better to err high (false negatives, which will be corrected by waiting a bit longer)
+            // than low (false positives, which cause "flapping").
+            arrivalIntervals.add(FailureDetector.INITIAL_VALUE_NANOS);
+        }
+        tLast = value;
+    }
+
+    double mean()
+    {
+        return arrivalIntervals.mean();
+    }
+
+    // see CASSANDRA-2597 for an explanation of the math at work here.
+    double phi(long tnow)
+    {
+        assert arrivalIntervals.mean() > 0 && tLast > 0; // should not be called before any samples arrive
+        long t = tnow - tLast;
+        lastReportedPhi = t / mean();
+        return lastReportedPhi;
+    }
+
+    public double getLastReportedPhi()
+    {
+        return lastReportedPhi;
+    }
+
+    public String toString()
+    {
+        return Arrays.toString(arrivalIntervals.getArrivalIntervals());
+    }
+
+    public long getLastInterval()
+    {
+        return arrivalIntervals.getLastInterval();
+    }
+}
